Used a NTSC framerate of 59.94hz

+++ Source/Core/Core/Core.cpp
@@ -622,7 +622,7 @@ void VideoThrottle()
 // depending on the framelimit set
 bool ShouldSkipFrame(int skipped)
 {
-	const u32 TargetFPS = (SConfig::GetInstance().m_Framelimit > 1)
+	const float TargetFPS = (SConfig::GetInstance().m_Framelimit > 1)
 		? (SConfig::GetInstance().m_Framelimit - 1) * 5
 		: VideoInterface::TargetRefreshRate;
 	const u32 frames = Common::AtomicLoad(s_drawn_frame);
